+++ BulletN.cpp
@@ -1,6 +1,6 @@
 #include "BulletN.h"
 
-CBulletN::CBulletN(float x, float y, int angle, bool friendly) :CBullet::CBullet(x, y, angle, friendly) {
+CBulletN::CBulletN(float x, float y, float angle, bool friendly) :CBullet::CBullet(x, y, angle, friendly) {
 	if (friendly == true) {
 		this->type = BULLET_ANI_NORMAL_BILL;
 	}
